Streamline envelope action menus
https://github.com/nextcloud/mail/issues/4256
